requirements.txt
elasticsearch==2.4.0
facebook-sdk==2.0.0
requests==2.11.1
urllib3==1.18
